Fix num_batches becoming float in Python 3
In Python 2, the division of two integers returns an integer. In Python 3 a division always returns a float. This results in problems when directly using the result of the division as num_batches in the generate_batches functions. Manually converting the results to integers keeps the same behavior as in Python 2 but resolves the error that occurs when using Python 3.